% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{covariates_matrix}
\alias{covariates_matrix}
\title{Extract Covariate Matrix (Z and X Matrix)}
\usage{
covariates_matrix(data, covariates, colname_unit)
}
\arguments{
\item{data}{A data frame containing the covariates.}

\item{covariates}{A character vector specifying which covariate columns to extract.}

\item{colname_unit}{A string specifying the name of the unit identifier column.}
}
\value{
A numeric matrix of covariates by unit.
}
\description{
This function specifies covariates columns by the unit column
and returns them as a numeric matrix, excluding the unit identifier column.
}
\examples{
test_data <- data.frame(
  unit = rep(c("A", "B"), each = 2),
  median_income = c(40, 50, 60, 70),
  median_age = c(10, 13, 15, 21),
  proportion = c(200, 234, 25, 32)
)
covariates <- c("median_income", "median_age", "proportion")
covariates_matrix(test_data, covariates, "unit")

}
