% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{generate_b_list}
\alias{generate_b_list}
\title{Generate List of B Vectors Summing to One}
\usage{
generate_b_list(step = 0.01, min_value = 0.01)
}
\arguments{
\item{step}{A numeric value indicating the increment step size. Default is \code{0.01}.}

\item{min_value}{A numeric value specifying the minimum threshold for each component. Default is \code{0.01}.}
}
\value{
A list of budgets, each represented as a list with elements \code{b_F}, \code{b_Z}, and \code{b_X}.
}
\description{
This function generates a list of combinations of a budgeting vector \eqn{(b_F, b_Z, b_X)} such that their sum equals 1.
}
\examples{
B_list <- generate_b_list(step = 0.01, min_value = 0.01)
length(B_list)  # View number of valid combinations
head(B_list)

}
