% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etts.R
\name{linear_equi_confounding_ett}
\alias{linear_equi_confounding_ett}
\title{Estimated Treatment Effect (ETT) using Linear Equi-Confounding}
\usage{
linear_equi_confounding_ett(Y_treated, Y_control, F_treated, F_control)
}
\arguments{
\item{Y_treated}{A numeric vector of outcomes for the treated group of target domain.}

\item{Y_control}{A numeric vector of outcomes for the control group of target domain.}

\item{F_treated}{A numeric vector of outcomes measurements for the treated group of reference domain.}

\item{F_control}{A numeric vector of outcomes measurements for the control group of reference domain.}
}
\value{
Estimated treatment effect (ETT) as a single numeric value.
}
\description{
This function computes the estimated treatment effect using a linear equi-confounding method.
The ETT is calculated to adjust the naive difference in outcomes between the treated and control units in the target domain by subtracting the corresponding difference in the reference domain.
}
\details{
\deqn{
\hat\psi^{eq1}=\left\{Y_1-F_1\right\}-\frac{1}{J}\sum_{i=2}^{J+1}\left\{Y_i- F_i\right\}.
}

where:
\itemize{
\item \eqn{\hat\psi^{eq1}} is the causal estimator using a linear equi-confounding method.
\item \eqn{Y_{1}} is the observed outcome for the treated unit of target domain
\item \eqn{Y_{i}} is the outcome for the control unit\eqn{i} of target domain.
\item \eqn{F_{1}} is the observed outcome for the treated unit of reference domain
\item \eqn{F_{i}} is the outcome for control for the control unit\eqn{i} of reference domain.
\item \eqn{J} is the number of control units.

}
}
