% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{synthetic_reference_plot_real_data}
\alias{synthetic_reference_plot_real_data}
\title{Plot Synthetic Reference Outcome Over Time}
\usage{
synthetic_reference_plot_real_data(
  F,
  w,
  t_max,
  i_max,
  J,
  title = NULL,
  x_label = "Time",
  y_label = "Outcome",
  legend_labels = c("Target Unit", "Synthetic Target Unit", "Control Units")
)
}
\arguments{
\item{F}{A matrix of outcomes of reference domain (rows = units, columns = time).}

\item{w}{A numeric vector of weights for synthetic control.}

\item{t_max}{Maximum number of time periods.}

\item{i_max}{Total number of units (including treated and controls).}

\item{J}{Number of control units.}

\item{title}{Optional plot title.}

\item{x_label}{Label for the x-axis. Default is "Time".}

\item{y_label}{Label for the y-axis. Default is "Outcomes in Target Domain".}

\item{legend_labels}{Character vector for custom legend labels. Default is "Target "}
}
\value{
A graph representing the comparison of outcomes for the observed target unit and synthetic target unit in the reference domain
}
\description{
This function creates a line plot comparing the treated target unit and its synthetic target unit
with other control units over time of the reference domain.
}
