% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{synthetic_target_plot_real_data}
\alias{synthetic_target_plot_real_data}
\title{Plot Synthetic Target Outcome Over Time}
\usage{
synthetic_target_plot_real_data(
  Y,
  w,
  s_max,
  i_max,
  J,
  title = NULL,
  x_label = "Time",
  y_label = "Outcomes in Target Domain",
  legend_labels = c("Target Unit", "Synthetic Target Unit", "Control Units")
)
}
\arguments{
\item{Y}{A matrix of outcomes for control units including treated unit in the first row.}

\item{w}{A numeric weight vector for constructing the synthetic control.}

\item{s_max}{Maximum number of time periods.}

\item{i_max}{Total number of units (including treated and controls).}

\item{J}{The number of control units.}

\item{title}{Optional plot title.}

\item{x_label}{Label for the x-axis. Default is "Time".}

\item{y_label}{Label for the y-axis. Default is "Outcomes in Target Domain".}

\item{legend_labels}{Character vector for custom legend labels. Default is "Target Unit", "Synthetic Target Unit", "Control Units"}
}
\value{
A graph representing the comparison of outcomes for the observed target unit and synthetic target unit in the target domain
}
\description{
This function creates a line plot comparing the treated target unit and its synthetic target unit
with other control units over time of the target domain.
}
